geolocation-API/watchPosition_permission_deny.https.html

<!DOCTYPE HTML>
<meta charset='utf-8'>
<title>Geolocation Test: watchPosition location access denied</title>
<link rel='help' href='http://www.w3.org/TR/geolocation-API/#watch-position'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='/resources/testdriver.js'></script>
<script src='/resources/testdriver-vendor.js'></script>
<script src='support.js'></script>
<script>
// Rewrite http://dev.w3.org/geo/api/test-suite/t.html?00062
promise_test(async t => {
  await test_driver.set_permission({name: 'geolocation'}, 'denied');

  let hasMethodReturned = false;
  let id = null;

  return new Promise((resolve, reject) => {
    id = geo.watchPosition(resolve, reject);
    hasMethodReturned = true;
  }).then(() => {
    assert_true(hasMethodReturned);
  }).catch(() => {
    assert_true(hasMethodReturned);

    // Rewrite http://dev.w3.org/geo/api/test-suite/t.html?00151
    assert_greater_than_equal(id, -2147483648);
    assert_less_than_equal(id, 2147483647);
  });
}, 'Check that watchPosition returns synchronously before any callbacks are invoked.');
</script>
